More improvements to fixed point shiny hunt.

+++ SerialPrograms/Source/PokemonLA/Inference/Objects/PokemonLA_FlagTracker.h
@@ -34,7 +34,7 @@ public:
 private:
     struct Sample{
         std::chrono::system_clock::time_point timestamp;
-        double distance;
+        int distance;
         double x;
         double y;
     };
